+++ src/base/classes/Version/ChangeLog.js
@@ -1,0 +1,21 @@
+/** 
+ > State Codes
+ * 0 - Not Released
+ * 1 - Released
+ * 2 - Waiting
+ * 3 - Being Applied
+ */
+
+export default [
+  {
+    version: "1.0.14",
+    build: "1000",
+    state: 0,
+
+    changes: {
+      fixes: [],
+      news: [],
+      improves: []
+    }
+  }
+];